requirements.txt
absl-py==2.3.0
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
narwhals==1.45.0
networkx==3.5
numba==0.61.2
numpy==1.26.4
onnxruntime==1.22.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==4.25.8
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
PyMatting==1.1.14
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
rembg==2.0.66
requests==2.32.4
rpds-py==0.25.1
scikit-image==0.25.2
scipy==1.16.0
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.2
streamlit==1.46.1
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.6.11
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
